delete confirm by user

+++ src/pages/MyCart/MyCart.jsx
@@ -1,24 +1,44 @@
-
+import Swal from 'sweetalert2'
+import withReactContent from 'sweetalert2-react-content'
 // import PropTypes from 'prop-types';
 
 import { useState } from "react";
 import { useLoaderData } from "react-router-dom";
 
 const MyCart = () => {
+    const MySwal = withReactContent(Swal)
     const cartsData = useLoaderData()
     const [cartData, setCartData] = useState(cartsData)
     const handleDeleteItems = id => {
-        fetch(`http://localhost:5000/cart/${id}`, {
-            method: "DELETE"
+        MySwal.fire({
+            title: 'Are you sure?',
+            text: "You won't be able to revert this!",
+            icon: 'warning',
+            showCancelButton: true,
+            confirmButtonColor: '#3085d6',
+            cancelButtonColor: '#d33',
+            confirmButtonText: 'Yes, delete it!'
+        }).then((result) => {
+            if (result.isConfirmed) {
+
+                fetch(`http://localhost:5000/cart/${id}`, {
+                    method: "DELETE"
+                })
+                    .then(res => res.json())
+                    .then(data => {
+                        if (data.deletedCount > 0) {
+                            console.log("data successfully  daleted ");
+                            MySwal.fire(
+                                'Deleted!',
+                                'Your file has been deleted.',
+                                'success'
+                            )
+                            const remainingUser = cartData.filter(user => user._id !== id)
+                            setCartData(remainingUser)
+                        }
+                    })
+            }
         })
-            .then(res => res.json())
-            .then(data => {
-                if (data.deletedCount > 0) {
-                    console.log("data successfully  daleted ");
-                    const remainingUser = cartData.filter(user => user._id !== id)
-                    setCartData(remainingUser)
-                }
-            })
     }
     return (
         <div className="overflow-x-auto mx-auto container">
